requirements.txt
numpy
slackclient
dynamixel-sdk
opencv-contrib-python==4.6.0.66
stag-python